correct api format and optimize code structure
