+++ models/logModel.go
@@ -1,0 +1,6 @@
+package models
+
+type EventLog struct {
+	Index uint8 `json:"index"`
+	Data string `json:"data"`
+}
